[add] add `exception/instruction_handler.rs`
