+++ src/trap/hypervisor_supervisor/exception.rs
@@ -1,5 +1,6 @@
 //! Trap VS-mode exception.
 
+mod instruction_handler;
 mod sbi_handler;
 
 use super::hstrap_exit;
@@ -13,10 +14,10 @@ use crate::memmap::HostPhysicalAddress;
 use crate::HYPERVISOR_DATA;
 
 use core::arch::asm;
-use raki::{Instruction, OpcodeKind};
+use raki::Instruction;
 use riscv::register::{
     scause::{self, Exception},
-    sepc, stval,
+    stval,
 };
 use sbi_handler::{sbi_base_handler, sbi_fwft_handler, sbi_rfnc_handler};
 
@@ -82,41 +83,9 @@ fn update_sepc_by_htinst_value(htinst_inst_value: usize, context: &mut guest::co
 /// Trap handler for exception
 #[allow(clippy::cast_possible_truncation, clippy::module_name_repetitions)]
 pub unsafe fn trap_exception(exception_cause: Exception) -> ! {
-    use crate::emulate_extension::zicfiss::ZICFISS_DATA;
-    use crate::emulate_extension::EmulateExtension;
-
     match exception_cause {
-        Exception::IllegalInstruction => {
-            let fault_inst_value = stval::read();
-            let fault_inst = Instruction::try_from(fault_inst_value)
-                .expect("decoding load fault instruction failed");
+        Exception::IllegalInstruction => instruction_handler::illegal_instruction(),
 
-            // emulate the instruction
-            match fault_inst.opc {
-                OpcodeKind::Zicfiss(_) => unsafe { ZICFISS_DATA.lock() }
-                    .get_mut()
-                    .unwrap()
-                    .instruction(&fault_inst),
-                OpcodeKind::Zicsr(_) => match fault_inst.rs2.unwrap() {
-                    // ssp
-                    0x11 => unsafe { ZICFISS_DATA.lock() }
-                        .get_mut()
-                        .unwrap()
-                        .csr(&fault_inst),
-                    unsupported_csr_num => {
-                        unimplemented!("unsupported CSRs: {unsupported_csr_num:#x}")
-                    }
-                },
-                _ => unimplemented!(
-                    "unsupported illegal instruction: {:#?}, at {:#x}",
-                    fault_inst,
-                    sepc::read()
-                ),
-            }
-
-            let mut context = unsafe { HYPERVISOR_DATA.lock().get().unwrap().guest().context };
-            context.update_sepc_by_inst(&fault_inst);
-        }
         Exception::SupervisorEnvCall => panic!("SupervisorEnvCall should be handled by M-mode"),
         // Enum not found in `riscv` crate.
         Exception::Unknown => match HvException::from(scause::read().code()) {
@@ -184,52 +153,7 @@ pub unsafe fn trap_exception(exception_cause: Exception) -> ! {
 
                 hs_forward_exception();
             }
-            HvException::VirtualInstruction => {
-                let fault_inst_value = stval::read();
-                let fault_inst = Instruction::try_from(fault_inst_value)
-                    .expect("decoding load fault instruction failed");
-                let mut context = unsafe { HYPERVISOR_DATA.lock() }
-                    .get()
-                    .unwrap()
-                    .guest()
-                    .context;
-
-                // emulate CSR set
-                match fault_inst.opc {
-                    OpcodeKind::Zicsr(_) => {
-                        match fault_inst.rs2.unwrap() {
-                            // senvcfg
-                            0x10a => {
-                                let mut read_from_csr_value: u64;
-                                unsafe {
-                                    asm!("csrr {0}, senvcfg", out(reg) read_from_csr_value);
-                                }
-
-                                let write_to_csr_value = context.xreg(fault_inst.rs1.unwrap());
-
-                                // update emulated CSR field.
-                                unsafe { ZICFISS_DATA.lock() }.get_mut().unwrap().csr_field(
-                                    &fault_inst,
-                                    write_to_csr_value,
-                                    &mut read_from_csr_value,
-                                );
-
-                                // commit result
-                                unsafe {
-                                    asm!("csrw senvcfg, {0}", in(reg) write_to_csr_value);
-                                }
-                                context.set_xreg(fault_inst.rd.unwrap(), read_from_csr_value);
-                            }
-                            unsupported_csr_num => {
-                                unimplemented!("unsupported CSRs: {unsupported_csr_num:#x}")
-                            }
-                        }
-                    }
-                    _ => unreachable!(),
-                }
-
-                context.update_sepc_by_inst(&fault_inst);
-            }
+            HvException::VirtualInstruction => instruction_handler::virtual_instruction(),
         },
         _ => hs_forward_exception(),
     }
